requirements.txt
facenet_pytorch==2.5.2
imutils==0.5.4
matplotlib==3.5.3
mtcnn==0.1.1
numpy==1.23.2
Pillow==9.2.0
setuptools==61.2.0
six==1.16.0
tensorflow<2.11
tensorflow_addons==0.17.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
opencv-python==4.6.0.66
scikit-learn
scipy